Add ExtendedSolver
