+++ dfpn-solver/src/impl_vec_table.rs
@@ -1,7 +1,7 @@
 use crate::{Table, U};
 
 pub struct VecTable {
-    table: Vec<(U, U)>,
+    table: Vec<Option<(U, U)>>,
     mask: usize,
     len: usize,
 }
@@ -9,7 +9,7 @@ pub struct VecTable {
 impl VecTable {
     pub fn new(bits: u32) -> Self {
         Self {
-            table: vec![(1, 1); 1 << bits],
+            table: vec![None; 1 << bits],
             mask: (1 << bits) - 1,
             len: 0,
         }
@@ -25,13 +25,13 @@ impl Default for VecTable {
 impl Table for VecTable {
     type T = usize;
     fn look_up_hash(&self, key: &Self::T) -> (U, U) {
-        self.table[*key & self.mask]
+        self.table[*key & self.mask].unwrap_or((1, 1))
     }
     fn put_in_hash(&mut self, key: Self::T, value: (U, U)) {
-        if self.look_up_hash(&key) == (1, 1) {
+        if self.table[key & self.mask].is_none() {
             self.len += 1;
         }
-        self.table[key & self.mask] = value;
+        self.table[key & self.mask] = Some(value);
     }
     fn len(&self) -> usize {
         self.len
